Moving exceptions to correct namespace
