+++ source/Octopus.Client.Tests/PublicSurfaceAreaFixture.ThePublicSurfaceAreaShouldNotRegress..NETFramework.approved.txt
@@ -1210,6 +1210,13 @@ Octopus.Client.Editors.Async
 }
 Octopus.Client.Exceptions
 {
+  class DefaultSpaceNotFoundException
+    ISerializable
+    _Exception
+    Exception
+  {
+    .ctor(Octopus.Client.Extensibility.IHaveSpaceResource)
+  }
   class OctopusDeserializationException
     ISerializable
     _Exception
@@ -1290,6 +1297,27 @@ Octopus.Client.Exceptions
     String ErrorMessage { get; }
     ReadOnlyCollection<String> Errors { get; }
     Octopus.Client.Exceptions.TDetails DetailsAs()
+  }
+  class ResourceSpaceDoesNotMatchRepositorySpaceException
+    ISerializable
+    _Exception
+    Exception
+  {
+    .ctor(Octopus.Client.Extensibility.IHaveSpaceResource, Octopus.Client.Model.SpaceResource)
+  }
+  class SpaceResourceIsIncompatibleWithSystemRepositoryException
+    ISerializable
+    _Exception
+    Exception
+  {
+    .ctor(Octopus.Client.Extensibility.IHaveSpaceResource)
+  }
+  class SystemResourceIsIncompatibleWithSpaceScopedRepositoryException
+    ISerializable
+    _Exception
+    Exception
+  {
+    .ctor(Octopus.Client.Extensibility.IResource)
   }
   class UnsupportedApiVersionException
     ISerializable
@@ -5800,13 +5828,6 @@ Octopus.Client.Operations
 }
 Octopus.Client.Repositories
 {
-  class DefaultSpaceNotFoundException
-    ISerializable
-    _Exception
-    Exception
-  {
-    .ctor(Octopus.Client.Extensibility.IHaveSpaceResource)
-  }
   interface IAccountRepository
     Octopus.Client.Repositories.IResourceRepository
     Octopus.Client.Repositories.ICreate<AccountResource>
@@ -6408,27 +6429,6 @@ Octopus.Client.Repositories
     .ctor(Octopus.Client.IOctopusRepository)
     Octopus.Client.Model.PerformanceConfigurationResource Get()
     Octopus.Client.Model.PerformanceConfigurationResource Modify(Octopus.Client.Model.PerformanceConfigurationResource)
-  }
-  class ResourceSpaceDoesNotMatchRepositorySpaceException
-    ISerializable
-    _Exception
-    Exception
-  {
-    .ctor(Octopus.Client.Extensibility.IHaveSpaceResource, Octopus.Client.Model.SpaceResource)
-  }
-  class SpaceResourceIsIncompatibleWithSystemRepositoryException
-    ISerializable
-    _Exception
-    Exception
-  {
-    .ctor(Octopus.Client.Extensibility.IHaveSpaceResource)
-  }
-  class SystemResourceIsIncompatibleWithSpaceScopedRepositoryException
-    ISerializable
-    _Exception
-    Exception
-  {
-    .ctor(Octopus.Client.Extensibility.IResource)
   }
 }
 Octopus.Client.Repositories.Async
