+++ source/Octopus.Client.Tests/Repositories/Async/BasicRepositoryFixture.cs
@@ -4,6 +4,7 @@ using FluentAssertions;
 using NSubstitute;
 using NUnit.Framework;
 using NUnit.Framework.Constraints;
+using Octopus.Client.Exceptions;
 using Octopus.Client.Extensibility;
 using Octopus.Client.Model;
 using Octopus.Client.Repositories;
